feat: fix the case where the browser disables audio for you
Browser's hate autoplaying videos so they do this thing where they
disable audio unless you click something on the page.

This fix adds a discrete message if notes are supposed to play, but
aren't playing because the browser has disabled the audio. It tells you
to just click anywhere and audio will be enabled again, and it seems to
work pretty well.

This commit also includes some basic styling fixes for mobile.