:fire: remove the speed-slider (temporarily?)
I want to do a better version of this in the future... one with a better
UI and using accurate musical tempos (with BPM values) like beepbox.co
does. I'm removing this one now so that I don't have to worry about
the backwards-compatibility of replacing this one with a future one.