:sound: fix: improve the sound quality by using samples
The samples were pulled from a beepbox recording of 7 pitches with
"Music Box 1" set as the instrument. This recording was sliced into
notes and loaded into a ToneJS sampler, that interpolates in order to
play the rest of the notes on the scale.

Currently, I'm playing the .mp3 notes, but I'm including the .wav ones
as well, in case I want to try those or use them as a fallback in the
future.