:bug: fix: comment out the twitter share button
These twitter buttons caused a race condition that was breaking
javascript on the page so I'm removing them temporarily to keep the
site running until I can put in a fix.

I think the issue had something to do with the load order of the twitter
javascript and the calling of window.twttr.widgets.load(this.element),
which assumes it's loaded on the page.