fix: add missing keys to the minify map
When people were creating 20-note songs, the new pitch values were being
stored in the URL in full, because they had no keys in the minify map.
This fixes the issue. Surprisingly, it doesn't break existing songs
because our key-renaming function will leave key names as-is, if
there's no corresponding value in the minify map.