Fix a bug where the paper length was incorrect when rendering new songs
When we made the "fix" in 44e76c7, we introduced a regression, where
clicking a song on the "featured" tab would rerender the song without
changing the number of "pages". This meant that if the new song was
longer, a bunch of the notes would be orphaned, below the paper.

This happened because only our setInitialSongLength function would
adjust the paper based on the songData, and we had restricted that to
only run on initial page load.

The solution was to set the song length on every render of this
component, and determine what to set it to based on the value being
passed to the render function.