Fix a bug where the paper divider "×" stayed visible as notes were added
This is a problem because we don't want users to be able to add notes to
a new page and then delete the page, leaving the notes orphaned.

The solution was to restore bubbling rerenders of this component, and
then to ensure that these rerenders don't set a new line-length var.

I tried allowing bubbling rerenders to run this.setNoteLineLengthVar,
(setting it to line length in the songData), but that resulted in
behaviors that felt unintuitive. For example:

  - Removing the last note from a page would delete that page
  - Adding a note to the final line of a page would add a new page

The solution in this commit avoids those issues