:bug: fix a regression where snap-to-grid can create overlapping notes.
Sometimes snap to grid will create a situation when the shadow note is
fully overlapping an existing note but clicking won't remove the
existing note because the cursor is outside of it. As a consequence two
overlapping notes would get placed.

Originally, I solved this by comparing the shadow note yPos with the
existing notes yPos's found in the store. If an identical yPos was found
in the store, then I'd remove the note instead of placing an overlapping
one.

This solution silently broke during my ToneJS refactor because I had
changed my stored yPos values to represent the center of the note and my
displayed yPos values to represent the top. This meant that my
comparison stopped matching and notes could overlap again.

I fixed it by making sure that I was comparing the center of the shadow
note to the stored value (which was the center of the existing note).