fix: prevent twitter share from accidentally sharing a stale url.
I noticed that shared songs were often missing a few characters from the
title. This was a result of the share button updating before the URL was
fully updated (which happened because the share button subscribed to state
changes before the URL did).

We could have reordered the initial subscriptions but the URL updates
being async would probably have preventing that from working. So
instead, we only populate the tweet right as a person is clicking it.
This has the nice benefit of removing an unnecessary subscription.