fix: prevent duplicate pubsub events from piling up when reconstructing components
Whenever we want a component to contain child components, we've used
'new' to create new component instances to replace the existing ones.
Unfortunately, the old subscribed render functions were still hanging
around in our pubsub system.

This update associates an ID with each renderTrigger subscribe
function. When it sees me trying to resubscribe the same ID to the same
event a second time, it replaces the original subscription with the new
one. I think it'll work and allow us to create more parent components
without requiring a huge refactor of our templating or events system.