fix: prevent audio from trailing scrolling
When pressing play, we used to tell scrolling and audio to play at the
same time, but the issue was that if the audio context hadn't been
started, there would be a delay because the ToneJS part wouldn't start
until the context was fully running. This would cause the playhead to
get ahead. Now, we check the audioContext, and wait for the resume()
promise to resolve before telling the part to start AND we trigger
scrolling from ToneJS, not from our appState.isPlaying state.

Finally, this also resolves an issue where hitting the space bar when
the audioContext hadn't started results in scrolling without audio. Now,
the user is alerted that they will need to click somewhere to start the
context.