feat: first pass for a PWA
This adds a manifest file and placeholder service worker (that doesn't
do anything). The lighthouse report for PWA passes, so the goal is to
see if our current code is enough to trigger an add-to-homescreen prompt.

It has proven difficult to test the add-to-homescreen prompt locally, so
I got frustrated and I'm just going to deploy it and test it in prod
because this is my project and I can do that if I want to, lol.