fix: remove index and unnecessary uniqueness constraint
We were running into an issue where we couldn't add songs to the
database when the data field was too long:

  "Values larger than 1/3 of a buffer page cannot be indexed..."

There are ways to hash the value and index the hash as unique, but it
seemed like a lot of hassle so I figured I'd just remove the uniqueness
constraint right now. If we need uniqueness in the future, we can bring
it back in other forms (like a check within the application instead of
the database).